requirements.txt
flask
flask-sqlalchemy
flask-migrate
requests
basilica
tweepy
scikit-learn
psycopg2-binary
gunicorn
python-dotenv
